Added fuctionality for frmAddService.
- Implemented Add Service functionality in frmAddService, including input validation, user confirmation, and integration with ServiceManager and ServiceRepo.
- Updated Service model to use float for pricePerHour to allow decimal values. Improved UI layout and error messaging consistency.

+++ Models/Service.cs
@@ -26,7 +26,7 @@ namespace DoggyDaycare.Models
 
         public Service(string name, 
                         string breedType, 
-                        int pricePerHour, 
+                        float pricePerHour, 
                         int maxCapacityPerSlot)
         {
             SetStatus("Active");
@@ -40,7 +40,7 @@ namespace DoggyDaycare.Models
                         string status, 
                         string name, 
                         string breedType, 
-                        int pricePerHour, 
+                        float pricePerHour, 
                         int maxCapacityPerSlot)
         {
             SetId(id);
